moved vector typedefs to their own files

+++ vertex/src/vertex/image/img/detail/base_type_defs.h
@@ -5,7 +5,7 @@
 namespace vx {
 namespace img {
 
-using size_type = size_t;
+using size_type = uint32_t;
 using byte_type = uint8_t;
 
 }
